+++ src/js/core/methods/index.js
@@ -59,8 +59,6 @@ import ApplyFlags from './ApplyFlags';
 import ApplySettings from './ApplySettings';
 import BackupDevice from './BackupDevice';
 import ChangePin from './ChangePin';
-import FirmwareErase from './FirmwareErase';
-import FirmwareUpload from './FirmwareUpload';
 import FirmwareUpdate from './FirmwareUpdate';
 import RecoveryDevice from './RecoveryDevice';
 
@@ -119,8 +117,6 @@ const classes: {[k: string]: any} = {
     'applySettings': ApplySettings,
     'backupDevice': BackupDevice,
     'changePin': ChangePin,
-    'firmwareErase': FirmwareErase,
-    'firmwareUpload': FirmwareUpload,
     'firmwareUpdate': FirmwareUpdate,
     'recoveryDevice': RecoveryDevice,
 };